strip down ui to keeps keys section, tenant and user logout
