+++ src/app/page.tsx
@@ -17,7 +17,7 @@ export default async function Home() {
   async function setTenantAndRedirect(tenant: string) {
     "use server";
     await sessionSetTenant(tenant);
-    return redirect("/v2/dashboard");
+    return redirect("/v2/settings/keys");
   }
 
   return (
