+++ src/components/header-section.tsx
@@ -1,0 +1,33 @@
+"use client";
+
+import { GalleryVerticalEnd } from "lucide-react";
+import * as React from "react";
+import { NavUser } from "@/components/nav-user";
+import { TeamSwitcher } from "@/components/team-switcher";
+import { useUser } from "@auth0/nextjs-auth0";
+
+const data = {
+  user: {
+    name: "shadcn",
+    email: "m@example.com",
+    avatar: "/avatars/shadcn.jpg",
+  },
+  teams: [
+    {
+      name: "Default Team",
+      logo: GalleryVerticalEnd,
+      plan: "Free",
+    },
+  ],
+};
+
+export function AppHeader({}: React.ComponentProps<"div">) {
+  const { user } = useUser();
+
+  return (
+    <div className="flex flex-row gap-4 justify-end w-[90%]">
+      <TeamSwitcher teams={data.teams} />
+      {user ? <NavUser user={user} /> : ""}
+    </div>
+  );
+}
